# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Atom Forge Construction API",
        "description": "API Documentation for Atom Forge Construction system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/admin/plans": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "Get all construction plans",
                "operationId": "7d783653d3dbf52629603bb7ad485601",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Plans"
                ],
                "summary": "Create a new construction plan",
                "operationId": "9ee53a32be843695d949b6f91ef61ce2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price_per_sqft"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Premium Plan"
                                    },
                                    "price_per_sqft": {
                                        "type": "number",
                                        "example": 3500
                                    },
                                    "features": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plan created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/plans/{plan}": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "Get plan details",
                "operationId": "cfc5cb5e51ab837d66cc1f78bafb31d0",
                "parameters": [
                    {
                        "name": "plan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Plans"
                ],
                "summary": "Update construction plan",
                "operationId": "45fe416921f0c8ff6e39333cec86d743",
                "parameters": [
                    {
                        "name": "plan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price_per_sqft": {
                                        "type": "number"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plan updated successfully"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plans"
                ],
                "summary": "Delete construction plan",
                "operationId": "1d07138c5dd3ebcc8a735761f4b09d3a",
                "parameters": [
                    {
                        "name": "plan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Plan deleted successfully"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}